fixup! move VideoSettings to cores
